requirements.txt
crewai>=0.28.0
crewai-tools>=0.3.0
pandas>=2.2.0
openpyxl>=3.1.2
pyyaml>=6.0.2
requests>=2.31.0
fastapi>=0.111.0
uvicorn>=0.30.0

langchain
langchain-openai
langchain-community
langchain-google-genai
langchain-groq
litellm